consolidate test package
