+++ app/testutil_test.go
@@ -8,8 +8,8 @@ import (
 	coretypes "github.com/tendermint/tendermint/types"
 
 	appns "github.com/celestiaorg/celestia-app/pkg/namespace"
-	"github.com/celestiaorg/celestia-app/testutil/blobfactory"
-	"github.com/celestiaorg/celestia-app/testutil/testfactory"
+	"github.com/celestiaorg/celestia-app/test/util/blobfactory"
+	"github.com/celestiaorg/celestia-app/test/util/testfactory"
 )
 
 func generateMixedTxs(normalTxCount, pfbCount, pfbSize int) ([][]byte, []tmproto.BlobTx) {
