+++ test/txsim/sequence.go
