+++ test/txsim/stake.go
