+++ test/util/testnode/sign.go
